test(firestore-bigquery-export): update e2e tests and cleanup
