--- firestore-bigquery-export/scripts/import/__tests__/emulator-params.env
@@ -1,2 +1,0 @@
-LOCATION=europe-west2
-PROJECT_ID=extensions-testing
