+++ firestore-bigquery-export/scripts/import/package.json
@@ -15,7 +15,7 @@
     "compile": "tsc",
     "import": "node ./lib/index.js",
     "prepare": "npm run build",
-    "test:local": "firebase ext:dev:emulators:exec ./node_modules/.bin/jest --test-params=./__tests__/emulator-params.env --project=extensions-testing --config=./__tests__/firebase.json"
+    "test:local": "jest"
   },
   "files": [
     "lib"
@@ -42,6 +42,7 @@
     "@types/workerpool": "^6.0.0",
     "axios": "^1.3.2",
     "chai": "^4.2.0",
+    "dotenv": "^16.3.1",
     "jest": "^29.3.1",
     "nanoid": "^4.0.0",
     "rimraf": "^2.6.3",
